+++ Dioxus.toml
@@ -29,7 +29,7 @@ watch_path = ["src"]
 [web.resource]
 
 # CSS style file
-style = ["/styles.css"]
+style = []
 
 # Javascript code file
 script = []
