+++ src/components/mod.rs
@@ -1,7 +1,9 @@
+pub(crate) mod app_container;
+pub(crate) mod app_nav;
 pub(crate) mod cabal_channel_list;
-pub(crate) mod cabal_nav_list;
 pub(crate) mod channel_posts;
 
+pub(crate) use self::app_container::AppContainer;
+pub(crate) use self::app_nav::AppNav;
 pub(crate) use self::cabal_channel_list::CabalChannelList;
-pub(crate) use self::cabal_nav_list::CabalNavList;
 pub(crate) use self::channel_posts::ChannelPosts;
